big changes!! API/Extenders is moved to PlugIns/Extenders Server IIS settings was changed to the global whist site at my own domain RoundStatus CurrentPlayer and LeadingPlayer was changed to the Enum created yesterday

+++ Server/PlugIn/Gamers/StrongerCardGamer.cs
@@ -15,7 +15,7 @@ namespace PlugIn.Gamers
             Suit? strongSuit = this.CurrentRoundStatus.Trump;
             Suit? turnSuit = null;
             if (this.CurrentRoundStatus.LeadingPlayer > 0)
-                turnSuit = this.CurrentRoundStatus.CurrentPlay[this.CurrentRoundStatus.LeadingPlayer].Value.Suit;
+                turnSuit = this.CurrentRoundStatus.CurrentPlay[(int)this.CurrentRoundStatus.LeadingPlayer].Value.Suit;
 
             Card chosen = default(Card);
             if (turnSuit.HasValue)
